+++ src/Order/Changelog.md
@@ -6,6 +6,7 @@
 ##### 2023-XX-YY
 
 - Added the `processing` order status enum value
+- Added the `ready_for_pickup` fulfillment status enum value
 
 ## 3.5.1
 ##### 2023-02-23